% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis-annotation.r
\name{annotated_y_axis}
\alias{annotated_y_axis}
\alias{annotated_x_axis}
\title{Annotations in the axis}
\usage{
annotated_y_axis(label, y, colour = waiver(), side = waiver(),
  tick = TRUE, digits = 2, parsed = FALSE, print_label = TRUE,
  print_value = TRUE, print_both = TRUE, sep = " = ", hjust = waiver(),
  vjust = waiver(), size = waiver(), fontface = waiver(),
  family = waiver(), rot = waiver(), ...)

annotated_x_axis(label, x, colour = waiver(), side = waiver(),
  tick = TRUE, digits = 2, parsed = FALSE, print_label = TRUE,
  print_value = TRUE, print_both = TRUE, sep = " = ", hjust = waiver(),
  vjust = waiver(), size = waiver(), fontface = waiver(),
  family = waiver(), rot = waiver(), ...)
}
\arguments{
\item{label}{Text to print}

\item{y}{Position on panel's y-scale for label}

\item{colour}{Colour of label}

\item{side}{left or right side to print annotation}

\item{...}{???}
}
\description{
Annotations in the axis
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(mpg, hp, colour=disp)) + geom_point()

l <- p + annotated_y_axis('hello', y=200, tick=TRUE)
l <- l + annotated_y_axis('nah', y=300, colour='blue') +
  annotated_y_axis('leftsidey', y=300, side='left') +
  annotated_y_axis("x^2 == .(y)", y=250, parsed=TRUE, vjust=1, tick=TRUE) +
  annotated_y_axis("bar(x) ~ yadayda == .(y)", y=150, parsed=TRUE, col='green', hjust=0, x=1)
(l <- p +   annotated_y_axis("x^2 == .(y)", y=250, parsed=TRUE, vjust=0.3, tick=TRUE) )
(l <- p +   annotated_y_axis("X == .(y)", y=250, parsed=TRUE, tick=TRUE) )

l
gt <- ggplot_gtable(ggplot_build(l))
data <- ggplot_build(l)
}
