% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_border.R
\name{element_box}
\alias{element_box}
\title{Box theme element}
\usage{
element_box(
  sides,
  left,
  top,
  right,
  bottom,
  fill = NULL,
  inherit.blank = FALSE,
  ...
)
}
\arguments{
\item{sides}{Any combination of \code{"top"},\code{"left"},\code{"right"},\code{"bottom"},
\code{"horizontal"},\code{"vertical"},\code{"L"},\code{"all"},\code{"none"},\code{TRUE},\code{FALSE}.}

\item{left, top, right, bottom}{A theme element (\code{element_line}); overruled by \code{sides}.}

\item{fill}{If set, causes a box to be drawn filled with this color.}

\item{inherit.blank}{Same as for \code{element_rect} and \code{element_line}.}

\item{...}{Additional parameters used for \code{element_line} for \code{sides}.}
}
\description{
Drop-in replacement for \code{element_rect}, where the four sides
can be specified individually.
}
\details{
Useful for specifying panels' borders, e.g. left-bottom lines, in facets
without modifying the axis-lines/-ticks.
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + theme(panel.border = element_box("L", color="blue"))
}
