% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-rep-lab.r, R/facet-wrap.r
\name{facet_rep_grid}
\alias{facet_rep_grid}
\alias{facet_rep_wrap}
\title{Repeat axis lines and labels across all facet panels}
\usage{
facet_rep_grid(..., repeat.tick.labels = FALSE)

facet_rep_wrap(..., repeat.tick.labels = FALSE)
}
\arguments{
\item{...}{Arguments used for \code{\link[ggplot2]{facet_grid}} or
\code{\link[ggplot2]{facet_wrap}}.}

\item{repeat.tick.labels}{Logical, removes the labels (numbering) on an inner
axes when \code{FALSE}.}
}
\description{
\code{\link[ggplot2]{facet_grid}} and \code{\link[ggplot2]{facet_wrap}}, but
with axis lines and labels preserved on all panels.
}
