% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-rep-lab.r
\name{facet_rep_wrap}
\alias{facet_rep_wrap}
\title{Wrap a 1d ribbon of panels into 2d}
\usage{
facet_rep_wrap(facets, shrink = TRUE, repeat.tick.labels = FALSE,
  .debug = FALSE, ...)
}
\arguments{
\item{facets}{Either a formula or character vector. Use either a
one sided formula, \code{~a + b}, or a character vector, \code{c("a", "b")}.}

\item{shrink}{If \code{TRUE}, will shrink scales to fit output of
statistics, not raw data. If \code{FALSE}, will be range of raw data
before statistical summary.}

\item{repeat.tick.labels}{Logical, removes the labels (numbering) on an axis when \code{FALSE}.}

\item{nrow, ncol}{Number of rows and columns.}

\item{scales}{should Scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"}).}

\item{strip.position}{By default, the labels are displayed on the top of
the plot. Using \code{strip.position} it is possible to place the labels on
either of the four sides by setting \code{strip.position = c("top",
"bottom", "left", "right")}}

\item{dir}{Direction: either "h" for horizontal, the default, or "v", for
vertical.}
}
\description{
\code{facet_wrap} wraps a 1d sequence of panels into 2d. This is generally
a better use of screen space than \code{\link{facet_grid}} because most
displays are roughly rectangular.
}

