% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\name{find_panel}
\alias{find_panel}
\title{Find panels in a gtable}
\usage{
find_panel(table)
}
\arguments{
\item{table}{A gtable}
}
\value{
A data.frame with some or all of the columns t(op), r(ight),
b(ottom), and l(eft)
}
\description{
These functions help detect the placement of panels in a gtable, if they are
named with "panel" in the beginning. \code{find_panel} returns the extend of
the panel area, while \code{panel_cols} and \code{panel_rows} returns the
columns and rows that contains panels respectively.
}
\details{
Taken directly from ggplot2/R/facet-.r
}
\keyword{internal}

