% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_utils.r
\name{get_panel_range}
\alias{get_panel_range}
\alias{get_panel_y_range}
\alias{get_panel_x_range}
\alias{get_panel_params}
\title{Version safe(r) method to get the y- and x-range from trained scales.}
\usage{
get_panel_y_range(layout, index = 1)

get_panel_x_range(layout, index = 1)

get_panel_params(layout, index = 1)
}
\arguments{
\item{layout}{\code{layout} part from ggplot_build}

\item{index}{Could be panel number?}
}
\description{
The names of the internal layout objects from \code{ggplot_build} changed
slightly.
}
