% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\name{if-not-null}
\alias{if-not-null}
\alias{\%||\%}
\title{Return if not null}
\usage{
a \%||\% b
}
\description{
Returns second argument if first argument is \code{NULL}.
}
\details{
Code taken from \file{ggplot2/R/utilities.r}.
}
\keyword{internal}
