% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable_show-.r
\name{is.small}
\alias{is.small}
\title{Is a given unit 'small'?}
\usage{
is.small(x)
}
\arguments{
\item{x}{A unit.}
}
\value{
Logical or \code{NA}.
}
\description{
Uses a holistic approach to determine whether a unit is 'small',
i.e. less than 1 cm, 1 line, 10 pt, or 0.4 in.
}
\details{
Based on arbitraily chosen definitions of 'small', this function can return
\code{TRUE} or \code{FALSE} if a unit is 'small'.

So far, less than 1 cm, 1 line, 10 pt, or 0.4 inches is defined as being
'small'.
Unresolved sizes, suchs as 'grobheight', 'grobwidth', or 'null' are not
small.
Units based on arithmetic, such as sum of multiple units,
are also \emph{not} small.
\code{NA}s are returned for undecided sizes.
}
