% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.r
\name{knit_print.data.frame}
\alias{knit_print.data.frame}
\alias{knit_print.tbl_df}
\alias{knit_print.grouped_df}
\title{knitr extension: Always use `kable` for data frames.}
\usage{
\method{knit_print}{data.frame}(x, options, ...)

\method{knit_print}{tbl_df}(x, ...)

\method{knit_print}{grouped_df}(x, ...)
}
\arguments{
\item{x}{an data frame or dplyr table object to be printed}

\item{options}{Current chunk options are passed through this argument.}

\item{...}{Ignored for now.}
}
\description{
Convenience function for working with R Notebooks that ensures data frames
(and dplyr tables) are printed with \code{\link[knitr]{knit_print}} while
allowing RStudio to render the data frame dynamically for inline display.
\strong{NB:} Automatically enables when package \code{splot} is loaded!
}
\details{
To disable these functions use \code{options(knit_print=FALSE)}. 
Set to \code{TRUE} to re-enable.

To supply \code{\link[knitr]{kable}} functions such as \code{caption} or 
\code{col.names}, set chunk options \code{kable.opts}:

\preformatted{
```{r kable.opts=list(caption='This is kable table caption.')}`
data.frame
```
}
}
\seealso{
\code{\link[knitr]{knit_print}}, \code{\link[knitr]{kable}}
}
\keyword{print}
