% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r, R/axis-annotation.r, R/coord-flex.r,
%   R/facet-rep-lab.r, R/facet-wrap.r, R/geom-pointline.r, R/geom-siderange.r
\docType{data}
\name{lemon-ggproto}
\alias{lemon-ggproto}
\alias{AAList}
\alias{CoordFlexCartesian}
\alias{CoordFlexFlipped}
\alias{CoordFlexFixed}
\alias{FacetGridRepeatLabels}
\alias{FacetWrapRepeatLabels}
\alias{GeomPointPath}
\alias{GeomPointLine}
\alias{GeomPointRangeLine}
\alias{GeomSideRange}
\title{ggproto classes used in lemon!}
\format{An object of class \code{AAList} (inherits from \code{ggproto}, \code{gg}) of length 6.}
\usage{
AAList
}
\arguments{
\item{side}{top, left, right, bottom, etc.}

\item{is.primary}{When TRUE, only get annotations whose side are set to that
side. When FALSE (default), also get those that are waiver.}
}
\description{
If you are creating a new geom, stat, position, or scale in another package,
you'll need to extend from \code{ggplot2::Geom}, \code{ggplot2::Stat},
\code{ggplot2::Position}, or \code{ggplot2::Scale}.
}
\seealso{
\code{\link[ggplot2]{ggproto}}, \code{\link[ggplot2]{ggplot2-ggproto}}
}
\keyword{datasets}
\keyword{internal}
