% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-rep-lab.r
\name{remove_labels_from_axis}
\alias{remove_labels_from_axis}
\title{Removes labels from axis grobs.}
\usage{
remove_labels_from_axis(axisgrob)
}
\arguments{
\item{axisgrob}{Grob with an axis.}
}
\description{
Called from FacetGridRepeatLabels.
}
\keyword{interal}
