% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.r
\name{reposition_legend}
\alias{reposition_legend}
\title{Reposition a legend onto a panel}
\usage{
reposition_legend(aplot, position = NULL, legend = NULL, panel = "panel",
  x = NULL, y = NULL, just = NULL, plot = TRUE)
}
\arguments{
\item{aplot}{a ggplot2 or gtable object.}

\item{position}{Where to place the legend in the panel. 
Overrules \code{x}, \code{y}, and \code{just} arguments.}

\item{legend}{The legend to place.}

\item{panel}{Name of panel in gtable.}

\item{x}{horisontal coordinate of legend, with 0 at left.}

\item{y}{vertical coordiante of legend, with 0 at bottom.}

\item{just}{'Anchor point' of legend; it is this point of the legend that is 
placed at the \code{x} and \code{y} coordinates.}

\item{plot}{Logical, when \code{TRUE} (default), draws plot with legend 
repositioned on a new page.}
}
\value{
gtable object, invisibly, with legend repositioned.
  Can be drawn with \code{\link[grid]{grid.draw}}.
}
\description{
Some descriptives under title.
}
\details{
To modify the look of the legend, use themes
}
\examples{
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
(d <- ggplot(dsamp, aes(carat, price)) +
 geom_point(aes(colour = clarity)))

reposition_legend(d + theme(legend.position='bottom'), 'bottom right')

# To change the orientation of the legend, use theme's descriptors.
reposition_legend(d + theme(legend.position='bottom'), 'top left')

# Use odd specifications, here offset the legend with half its height from the bottom.
reposition_legend(d + theme(legend.position='bottom'), x=0.3, y=0, just=c(0, -0.5))
}

