% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r, R/coord-flex.r, R/facet-rep-lab.r,
%   R/facet-wrap.r
\docType{data}
\name{splot-ggproto}
\alias{splot-ggproto}
\alias{CoordFlexCartesian}
\alias{CoordFlexFlipped}
\alias{CoordFlexFixed}
\alias{FacetGridRepeatLabels}
\alias{FacetWrapRepeatLabels}
\title{ggproto classes used in splot!}
\description{
If you are creating a new geom, stat, position, or scale in another package,
you'll need to extend from \code{ggplot2::Geom}, \code{ggplot2::Stat},
\code{ggplot2::Position}, or \code{ggplot2::Scale}.
}
\seealso{
\code{\link[ggplot2]{ggproto}}, \code{\link[ggplot2]{ggplot2-ggproto}}
}
\keyword{internal}
