% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\name{waiver}
\alias{\%|W|\%}
\alias{is.waive}
\alias{waiver}
\title{A waiver object.}
\usage{
waiver()

is.waive(x)

a \%|W|\% b
}
\arguments{
\item{x}{The object to inquery is a \code{waiver}.}
}
\description{
A waiver is a "flag" object, similar to \code{NULL}, that indicates the
calling function should just use the default value.  It is used in certain
functions to distinguish between displaying nothing (\code{NULL}) and
displaying a default value calculated elsewhere (\code{waiver()})
}
\details{
Code taken from \file{ggplot2/R/utilities.r}.
}
\references{
ggplot2
}
\seealso{
\code{\link[ggplot2]{waiver}}
}
\keyword{internal}

