% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splot.r
\docType{package}
\name{splot}
\alias{splot}
\alias{splot-package}
\title{Stefan's plot package}
\source{
\url{https://github.com/stefanedwards/splot}
}
\description{
Just another ggplot2 extension.
}
\section{Functions for axis}{


See \code{\link{coord_capped_cart}} and \code{\link{coord_flex_cart}}. 
The latter is a shorthand version of the former. 
It automatically uses \code{\link{capped_horisontal}} and 
\code{\link{capped_vertical}}, but both accepts these as well as 
\code{\link{brackets_horisontal}} and \code{\link{brackets_vertical}}.
}
\author{
Stefan McKinnon Edwards <sme@iysik.com>
}

