% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.r
\name{g_legend}
\alias{g_legend}
\title{Extract ggplot legends}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{ggplot2 object}
}
\value{
gtable (grob) object. Draw with \code{\link[grid]{grid.draw}}.
}
\description{
Extracts the legend ('guide-box') from a ggplot2 object.
}
\details{
The extraction is applied \emph{after} the plot is trained and themes are 
applied. Modifying the legend is easiest by applying themes etc.
to the ggplot2 object, before calling \code{g_legend}.
}
\examples{
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
(d <- ggplot(dsamp, aes(carat, price)) +
 geom_point(aes(colour = clarity)))
 
legend <- g_legend(d)
grid.newpage()
grid.draw(legend)
}
\author{
Stack Overflow
}

