% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r, R/coord-flex.r, R/facet-rep-lab.r
\docType{data}
\name{ggplot2-ggproto}
\alias{CoordFlexCartesian}
\alias{CoordFlexFixed}
\alias{CoordFlexFlipped}
\alias{FacetGridRepeatLabels}
\alias{FacetWrapRepeatLabels}
\alias{ggplot2-ggproto}
\title{Base ggproto classes for ggplot2}
\description{
If you are creating a new geom, stat, position, or scale in another package,
you'll need to extend from \code{ggplot2::Geom}, \code{ggplot2::Stat},
\code{ggplot2::Position}, or \code{ggplot2::Scale}.
}
\seealso{
\code{\link[ggplot2]{ggproto}}
}
\keyword{datasets}
\keyword{internal}

