% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r, R/guides-axis.r
\name{waiver}
\alias{waiver}
\alias{is.waive}
\alias{\%|W|\%}
\alias{if-not-null}
\alias{\%||\%}
\alias{guide_axis}
\title{The waivers are from \file{ggplot2/R/utilities.r}.}
\usage{
waiver()

is.waive(x)

a \%|W|\% b

a \%||\% b

guide_axis(at, labels, position = "right", theme)
}
\arguments{
\item{x}{The object to inquery is a \code{waiver}.}
}
\description{
The waivers are from \file{ggplot2/R/utilities.r}.

From R/guide-axes.r
}
\seealso{
\code{\link[ggplot2]{waiver}}
}
\keyword{internal}
