% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.r
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Share a legend between multiple plots}
\source{
Originally brought to you by \href{http://rpubs.com/sjackman}{Shaun Jackman} 
  (\href{http://rpubs.com/sjackman/grid_arrange_shared_legend}{original}),
  and further improved by \code{baptiste} at
  \url{https://github.com/tidyverse/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs}
}
\usage{
grid_arrange_shared_legend(..., ncol = length(list(...)), nrow = 1,
  position = c("bottom", "right"), plot = TRUE)
}
\arguments{
\item{...}{ggplot2 objects. Their legends are automatically hidden. 
The legend is taken from the first argument.}

\item{ncol}{Integer, number of columns to arrange plots in.}

\item{nrow}{Integer, number of rows to arrange plots in.}

\item{position}{'bottom' or 'right' for positioning legend.}

\item{plot}{Logical, when \code{TRUE} (default), draws combined plot on a
new page.}
}
\value{
gtable of combined plot, invisibly.
}
\description{
Extract legend, combines plots using \code{\link[gridExtra]{arrangeGrob}},
and places legend underneath.
}
\examples{
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
p1 <- qplot(carat, price, data = dsamp, colour = clarity)
p2 <- qplot(cut, price, data = dsamp, colour = clarity)
p3 <- qplot(color, price, data = dsamp, colour = clarity)
p4 <- qplot(depth, price, data = dsamp, colour = clarity)
grid_arrange_shared_legend(p1, p2, p3, p4, ncol = 4, nrow = 1)
grid_arrange_shared_legend(p1, p2, p3, p4, ncol = 2, nrow = 2)
}
\seealso{
\code{\link{g_legend}}
}

