% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemon-plot.r
\name{as.lemon_plot}
\alias{as.lemon_plot}
\alias{ggplot_build.lemon_plot}
\alias{ggplot_gtable.built_lemon}
\alias{ggplot_add.lemon_scale}
\title{Lemon plots; ggplots with extended functionality.}
\usage{
as.lemon_plot(plot)

\method{ggplot_build}{lemon_plot}(plot)

\method{ggplot_gtable}{built_lemon}(data)

\method{ggplot_add}{lemon_scale}(object, plot, object_name)
}
\arguments{
\item{plot}{ggplot object}
}
\description{
Lemon plot basically hijacks ggplot2's building process and allows us
to change some things.
}
\keyword{internal}
