% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemon_print.r
\name{lemon_print_data_frame}
\alias{lemon_print_data_frame}
\alias{lemon_print_table}
\alias{load_lemon_print}
\title{knitr extension: Always use `kable` for data frames.}
\usage{
lemon_print_data_frame(x, options, ...)

lemon_print_table(x, options = list(), ...)

load_lemon_print(ignore = FALSE, safe = FALSE)
}
\arguments{
\item{x}{an data frame or dplyr table object to be printed}

\item{options}{Current chunk options are passed through this argument.}

\item{...}{Ignored for now.}

\item{ignore}{Logical; when \code{TRUE}, print warning when overwriting 
pre-existing names in \code{\link[base]{.GlobalEnv}}.}

\item{safe}{Logical; when \code{TRUE}, does not overwrite pre-existing names
in \code{\link[base]{.GlobalEnv}}.}
}
\description{
Convenience function for working with R Notebooks that ensures data frames
(and dplyr tables) are printed with \code{\link[knitr]{kable}} while
allowing RStudio to render the data frame dynamically for inline display.
}
\details{
\code{load_lemon_print} pushes \code{lemon_print} functions into relevant
S3 functions (\code{knit_print}) in the global namespace, potentially
overriding pre-existing functions.
}
\section{Knitr chunks}{


To supply \code{\link[knitr]{kable}} functions such as \code{caption} or
\code{col.names}, set chunk options \code{kable.opts}:

\preformatted{
```{r kable.opts=list(caption='This is kable table caption.')}`
data.frame
```
}

To disable, temporarily, specify chunk option:
\preformatted{
```{r render=normal_print}`
data.frame
```
}
}

\seealso{
\code{\link[knitr]{knit_print}}, \code{\link[knitr]{kable}}
}
