% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-rep-lab.r
\name{remove_labels_from_axis}
\alias{remove_labels_from_axis}
\title{Removes labels from axis grobs.}
\usage{
remove_labels_from_axis(axisgrob, direction = c("horizontal", "vertical"))
}
\arguments{
\item{axisgrob}{Grob with an axis.}

\item{direction}{Whether the axis is horizontal or vertical.}
}
\description{
Called from FacetGridRepeatLabels.
}
\keyword{interal}
